Use new SocketTime replacing Instant
Instant is a complicated type as it's opaque, and only useful
with Duration. It cannot be created from a value, and it cannot
be serialized and deserialized, and works poorly when interfacing
C++. It typically represents the current time.

This change introduces a type called SocketTime that represent a
point in time relative to the socket's creation time. It's a strong
type, wrapping an u64 representing nanoseconds since the socket was 
created.

This simplifies interfacing with C++, and overall simplifies the API.